+++ tests/integration/requestAuthPlugin.int.js
@@ -1,4 +1,5 @@
 import Iron from 'iron';
+import { MEMBER_COOKIE } from '../../src/util/cookieUtils';
 import { getServer } from '../../src/util/testUtils';
 import requestAuthPlugin from '../../src/plugins/requestAuthPlugin';
 
@@ -20,13 +21,16 @@ const makeMockFetchResponse = responseObj =>
 
 const random32 = 'asdfasdfasdfasdfasdfasdfasdfasdf';
 const config = {
-	API_HOST: 'www.api.meetup.com',
-	API_TIMEOUT: 10,
-	CSRF_SECRET: random32,
-	COOKIE_ENCRYPT_SECRET: random32,
-	OAUTH_AUTH_URL: 'https://secure.dev.meetup.com/oauth2/authorize',
-	OAUTH_ACCESS_URL: 'https://secure.dev.meetup.com/oauth2/access',
+	api: {
+		protocol: 'https',
+		host: 'www.api.meetup.com',
+		timeout: 10,
+	},
+	csrf_secret: random32,
+	cookie_encrypt_secret: random32,
 	oauth: {
+		auth_url: 'https://secure.dev.meetup.com/oauth2/authorize',
+		access_url: 'https://secure.dev.meetup.com/oauth2/access',
 		key: random32,
 		secret: random32,
 	},
@@ -42,12 +46,12 @@ const expectedResponse = 'barfoo';
 
 const testAuth = (cookies, test, makeRequest = cookieRequest) => {
 	spyOn(global, 'fetch').and.callFake((url, opts) => {
-		if (url.includes(config.OAUTH_AUTH_URL)) {
+		if (url.includes(config.oauth.auth_url)) {
 			return makeMockFetchResponse({
 				code: 'foo',
 			});
 		}
-		if (url.includes(config.OAUTH_ACCESS_URL)) {
+		if (url.includes(config.oauth.access_url)) {
 			return makeMockFetchResponse({
 				oauth_token: expectedOauthToken,
 				refresh_token: 'whatever',
@@ -60,7 +64,7 @@ const testAuth = (cookies, test, makeRequest = cookieRequest) => {
 		path: '/foo',
 		handler: (request, reply) => reply(expectedResponse),
 	};
-	const server = getServer({}, config);
+	const server = getServer(config);
 	return server
 		.register({
 			register: requestAuthPlugin,
@@ -76,10 +80,10 @@ const testAuth = (cookies, test, makeRequest = cookieRequest) => {
 describe('logged-in member state', () => {
 	// logged-in auth provides MEETUP_MEMBER
 	it('Passes MEETUP_MEMBER value as request credentials', () => {
-		const cookies = { MEETUP_MEMBER: 'foo' };
+		const cookies = { [MEMBER_COOKIE]: 'foo' };
 		const test = response => {
 			expect(response.payload).toEqual(expectedResponse);
-			expect(response.request.auth.credentials).toBe(cookies.MEETUP_MEMBER);
+			expect(response.request.auth.credentials).toBe(cookies[MEMBER_COOKIE]);
 		};
 		return testAuth(cookies, test);
 	});
@@ -94,10 +98,10 @@ describe('logged-in member state', () => {
 					.join('; '),
 			},
 		});
-		const cookies = { MEETUP_MEMBER: 'foo' };
+		const cookies = { [MEMBER_COOKIE]: 'foo' };
 		const test = response => {
 			expect(response.payload).toEqual(expectedResponse);
-			expect(response.request.state.MEETUP_MEMBER).toBeNull();
+			expect(response.request.state[MEMBER_COOKIE]).toBeNull();
 		};
 		return testAuth(cookies, test, makeLogoutRequest);
 	});